+++ src/common/strlib.c
@@ -1058,7 +1058,7 @@ static int StringBuf_AppendStr(StringBuf *self, const char *str)
 	if( needed >= available ) {
 		// not enough space, expand the buffer (minimum expansion = 1024)
 		size_t off = (self->ptr_ - self->buf_);
-		self->max_ += max(needed, 1024);
+		self->max_ += max(needed, (size_t)1024);
 		self->buf_ = (char *)aRealloc(self->buf_, self->max_ + 1);
 		self->ptr_ = self->buf_ + off;
 	}
